+++ cmd/podman/machine/platform.go
@@ -29,7 +29,7 @@ func GetSystemProvider() (machine.VirtProvider, error) {
 	logrus.Debugf("Using Podman machine with `%s` virtualization provider", resolvedVMType.String())
 	switch resolvedVMType {
 	case machine.QemuVirt:
-		return qemu.GetVirtualizationProvider(), nil
+		return qemu.VirtualizationProvider(), nil
 	default:
 		return nil, fmt.Errorf("unsupported virtualization provider: `%s`", resolvedVMType.String())
 	}
